fix(usb): xhci_core UB on 32-bit target; driver builds op-reg image in RAM
Fix 1: widen Link-TRB param_hi shift from (uintptr_t) to (uint64_t)
before >> 32 to avoid undefined behavior on the M55 32-bit target
(C11 §6.5.7p3 forbids shift by operand width).

Fix 2: xhci_init_sequence now writes a RAM shadow op_image instead of
the real MMIO op-reg block. The op-reg pointer to (cfg->base+0x20) is
removed; data->op_image receives the validated core output. A
consolidated TODO(aen401-bench) block describes the full bench sequence:
CAPLENGTH read, DWC3 soft-reset, HCRST + CNR poll, then volatile
copy of DCBAAP/CRCR/CONFIG from op_image, and USBCMD.R/S at _enable.

+++ zephyr/drivers/usb/uhc/xhci_core.c
@@ -18,7 +18,7 @@ void xhci_ring_init(struct xhci_ring *ring, struct xhci_trb *seg, uint32_t size)
 	 * cycle bit is set to the producer cycle as the ring crosses it.
 	 * Ring Segment Base is a 64-bit address (spec §6.4.4.1). */
 	seg[size - 1u].param_lo = (uint32_t)(uintptr_t)seg;
-	seg[size - 1u].param_hi = (uint32_t)((uintptr_t)seg >> 32);
+	seg[size - 1u].param_hi = (uint32_t)((uint64_t)(uintptr_t)seg >> 32);
 	seg[size - 1u].control  = XHCI_TRB_TYPE(XHCI_TRB_TYPE_LINK) | XHCI_TRB_LINK_TC;
 }
